Modularize LuaSF source layout
