+++ src/luasf/probability.lua
@@ -1,0 +1,13 @@
+--[[
+Probability helpers placeholder.
+
+This module is intentionally small for now. Future versions may include
+factorial, combinations, permutations, and related probability helpers.
+
+Those helpers should be added carefully to avoid overflow and to keep LuaSF
+focused on lightweight statistics, probability, and simulation utilities.
+]]
+
+local M = {}
+
+return M
